move snapshoter test to existing test rather than separate
